apps/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
SQLAlchemy>=2.0.29
alembic>=1.13.1
psycopg2-binary>=2.9.9
celery[redis]>=5.3.6
redis>=5.0.3
python-multipart>=0.0.9
boto3>=1.34.84
PyMuPDF>=1.24.1
python-docx>=1.1.0
beautifulsoup4>=4.12.3
authlib>=1.3.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
